Remove all routes in run.py
